fixed size_t error

+++ main.c
@@ -26,7 +26,7 @@ int terminationStatus = 0; // last termination code
 
 void newChild(){
     char *argsToRun[numCmds+1];
-    for(size_t i = 0; i < numCmds+1; i++){
+    for(int i = 0; i < numCmds+1; i++){
         argsToRun[i] = commandArgs[i];
     }
     numCmds = 0;
